Cherry-pick `_` methods
This is to shrink the codebase down more so it creates a smaller UMD
bundle.

If further shrinking is required, remove `_` dependencies and write your
own less general, project specific `_` methods. For example,
`_.isEmpty()` currently is only needed for `String`s and `Object`s.
Also, underscore.string methods coerce objects into strings and that may
not be required.